webabc/requirement.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
